requirements.txt
numpy==1.21.0
Pillow==8.3.0
PyQt5==5.15.4
PyQt5-Qt5==5.15.2
PyQt5-sip==12.9.0
pyinstaller==4.3
